refactor(arch): break config ↔ utils circular dependency
Extract DATA_DIR and get_data_dir to new leaf module teatree.paths
(no dependencies). Update all consumers to import from paths directly.

redis_container: remove redis_db_count() — callers now pass db_count
from config, keeping utils free of config imports.

tach.toml: utils no longer depends on config; paths added as leaf.
