+++ src/teatree/cli/overlay.py
@@ -202,7 +202,7 @@ class OverlayAppBuilder:
         @overlay_app.command()
         def resetdb() -> None:
             """Drop the SQLite database and re-run all migrations."""
-            from teatree.config import get_data_dir  # noqa: PLC0415
+            from teatree.paths import get_data_dir  # noqa: PLC0415
 
             db_path = get_data_dir(overlay_name) / "db.sqlite3"
             if db_path.exists():
