+++ src/teatree/core/skill_cache.py
@@ -16,8 +16,8 @@ import operator
 from pathlib import Path
 
 import teatree
-from teatree.config import DATA_DIR
 from teatree.core.overlay_loader import get_overlay
+from teatree.paths import DATA_DIR
 from teatree.skill_deps import resolve_all
 from teatree.skill_schema import validate_skill_md
 from teatree.trigger_parser import parse_triggers as _parse_triggers
